requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==9.0.1
click==8.0.3
flask-cors==3.0.10
flask-mongoalchemy==0.7.1
flask-restful==0.3.9
flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1
mongoalchemy==0.21
pyjwt==2.3.0
pymongo==3.12.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
typing-extensions==3.10.0.2
werkzeug==2.0.2
zipp==3.6.0
